backend/requirements.txt
asgiref==3.9.1
attrs==25.4.0
channels==4.3.2
channels_redis==4.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-simple-history==3.10.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
factory_boy==3.3.3
Faker==38.2.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
reportlab==4.3.1
